Fix clippy warnings and add unit tests for astgrep-core
Fix &PathBuf->&Path, Default impls, rename not()->pattern_not(). Add ~100+ unit tests covering config, execution, patterns, models, types, and error modules.

Ultraworked with [Sisyphus](https://github.com/code-yeongyu/oh-my-openagent)

+++ crates/astgrep-core/src/models/test_asset.rs
@@ -246,7 +246,7 @@ impl TestAsset {
         self.metadata.file_size.or_else(|| {
             std::fs::metadata(&self.current_path)
                 .ok()
-                .map(|m| m.len() as u64)
+                .map(|m| m.len())
         })
     }
